+++ src/main/java/gg/archipelago/APClient/parts/Game.java
@@ -11,8 +11,8 @@ public class Game implements Serializable {
     int version;
 
     @SerializedName("item_name_to_id")
-    HashMap<String,Integer> itemNameToId = new HashMap<>();
+    HashMap<String,Long> itemNameToId = new HashMap<>();
 
     @SerializedName("location_name_to_id")
-    HashMap<String,Integer> locationNameToId = new HashMap<>();
+    HashMap<String,Long> locationNameToId = new HashMap<>();
 }
